torch.testing allclose -> close
Right now the functions have divergent names with one postfixed `_equal` and the other `_allclose`. I've opted to use `_(equal|close)` over `_all(equal|close)` think it is a reasonable assumption that all values need to be equal or close for this pass even without explicitly naming the function this way.

ghstack-source-id: 9c584502c38970d660faefc1b449558a7fcd127f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54781